Update DiffDriveROSserial.ino
Made pwm values doubles for min max comparison.

+++ src/scripts/DiffDriveROSserial/DiffDriveROSserial.ino
@@ -19,7 +19,7 @@ std_msgs::UInt64 enc_lf_msg;
 std_msgs::UInt64 enc_lb_msg;
 std_msgs::UInt64 enc_rf_msg;
 std_msgs::UInt64 enc_rb_msg;
-long lpwm, rpwm;
+double lpwm, rpwm;
 int motorPWM[2];
 int forPins[2] = { 16, 18 };  // 16 right forward, 18 left forward
 int bacPins[2] = { 17, 19 };  // 17 right backward, 19 right backward
